+++ source/extensions/isaacsim.robot_setup.assembler/docs/CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
+
+## [3.0.0] - 2025-05-14
+### Changed
+- Update to use new robot schema
+- Updated robot assembler API
+- Updated UI
+- Updated test cases
+### Added
+- Apply changes on Payloads and configurable with variants when assembling directly on source base robot.
+### Removed
+- Deprecated Script generation for robot assembler - new API should be straightforward to use.
 ## [2.1.13] - 2025-05-10
 ### Changed
 - Enable FSD in test settings